Added transaction settings component to other pages
