+++ src/state/user/reducer.tsx
@@ -3,10 +3,14 @@ import { INITIAL_ALLOWED_SLIPPAGE } from '../../utils/constants'
 import {
   updateUserSlippageTolerance,
 } from './actions'
+import { updateVersion } from '../global/actions'
 
 const currentTimestamp = () => new Date().getTime()
 
 export interface UserState {
+  // the timestamp of the last updateVersion action
+  lastUpdateVersionTimestamp?: number
+  
   // user defined slippage tolerance in bips, used in all txns
   userSlippageTolerance: number
 
@@ -23,6 +27,14 @@ export const initialState: UserState = {
 
 export default createReducer(initialState, (builder) =>
   builder
+    .addCase(updateVersion, (state) => {
+      // slippage isnt being tracked in local storage, reset to default
+      // noinspection SuspiciousTypeOfGuard
+      if (typeof state.userSlippageTolerance !== 'number') {
+        state.userSlippageTolerance = INITIAL_ALLOWED_SLIPPAGE
+      }
+      state.lastUpdateVersionTimestamp = currentTimestamp()
+    })
     .addCase(updateUserSlippageTolerance, (state, action) => {
       state.userSlippageTolerance = action.payload.userSlippageTolerance
       state.timestamp = currentTimestamp()
